extract the pinner to go-ipfs-pinner and dagutils into go-merkledag
