+++ Rules.mk
@@ -48,9 +48,6 @@ ifneq ($(filter coverage% clean distclean test/unit/gotest.junit.xml,$(MAKECMDGO
 	include $(dir)/Rules.mk
 endif
 
-dir := pin/internal/pb
-include $(dir)/Rules.mk
-
 # -------------------- #
 #   universal rules    #
 # -------------------- #
