--- pin/internal/pb/Rules.mk
@@ -1,8 +1,0 @@
-include mk/header.mk
-
-PB_$(d) = $(wildcard $(d)/*.proto)
-TGTS_$(d) = $(PB_$(d):.proto=.pb.go)
-
-#DEPS_GO += $(TGTS_$(d))
-
-include mk/footer.mk
