--- pin/internal/pb/doc.go
@@ -1,3 +1,0 @@
-package pb
-
-//go:generate protoc --gogo_out=. header.proto
